covering the get-first function
