+++ lib/get-primary-value.js
@@ -1,13 +1,6 @@
 'use strict';
 
-function getFirst (arr) {
-  if (!arr[0]) {
-    return null;
-  }
-  return arr[0].type !== 'note'
-       ? arr[0].value
-       : ((arr[1] || {}).value || null);
-}
+const getFirst = require('./get-first.js');
 
 module.exports = function getPrimaryValue (array) {
   var primary;
